[Code] add service worker
