+++ src/index.tsx
@@ -6,7 +6,6 @@ import React from "react";
 import { createRoot } from "react-dom/client";
 import { createBrowserRouter, RouterProvider } from "react-router-dom";
 import routes from "routes";
-import * as serviceWorker from "./serviceWorker";
 
 const router = createBrowserRouter(routes);
 
@@ -22,4 +21,4 @@ root.render(
 // If you want your app to work offline and load faster, you can change
 // unregister() to register() below. Note this comes with some pitfalls.
 // Learn more about service workers: https://bit.ly/CRA-PWA
-serviceWorker.unregister();
+serviceWorker.register();
